+++ bootstrap/server.go
@@ -1,0 +1,46 @@
+package main
+
+import (
+	"flag"
+	"net/http"
+	_ "net/http/pprof"
+
+	"github.com/AVENTER-UG/util"
+	"github.com/gorilla/mux"
+	"github.com/sirupsen/logrus"
+)
+
+var MinVersion string
+
+// Commands is the main function of this package
+func Commands() *mux.Router {
+	// Connect with database
+
+	rtr := mux.NewRouter()
+	rtr.HandleFunc("/versions", APIVersions).Methods("GET")
+
+	return rtr
+}
+
+// APIVersions give out a list of Versions
+func APIVersions(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("Api-Service", "-")
+	w.Write([]byte("/api/user/v0"))
+}
+
+func main() {
+	util.SetLogging("INFO", false, "GO-K3S-API")
+
+	bind := flag.String("bind", "0.0.0.0", "The IP address to bind")
+	port := flag.String("port", "10422", "The port to listen")
+
+	logrus.Println("GO-K3S-API build"+MinVersion, *bind, *port)
+
+	http.Handle("/", Commands())
+
+	if err := http.ListenAndServe(*bind+":"+*port, nil); err != nil {
+		logrus.Fatalln("ListenAndServe: ", err)
+	}
+
+}